feat: update gif component
